working down the errors, now have to handle clone_shallow in a way that lets me drop the borrow first

+++ src/action.rs
@@ -5,8 +5,8 @@ pub enum Action {
     Select(Expr),
     OpSet(Op, Expr, Expr),
     Set(Expr, Expr),
-    AddItem(i32, String),
-    RemItem(i32, String),
+    AddItem(isize, String),
+    RemItem(isize, String),
     Return(Expr),
     Display(Expr),
 }
